Fixed merge errors & unit tests
