+++ app/post.go
@@ -7,6 +7,8 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"github.com/cosmos/gaia/v18/ante"
 )
 
 // PostHandlerOptions are the options required for constructing a FeeMarket PostHandler.
@@ -19,6 +21,10 @@ type PostHandlerOptions struct {
 
 // NewPostHandler returns a PostHandler chain with the fee deduct decorator.
 func NewPostHandler(options PostHandlerOptions) (sdk.PostHandler, error) {
+	if !ante.UseFeeMarketDecorator {
+		return nil, nil
+	}
+
 	if options.AccountKeeper == nil {
 		return nil, errorsmod.Wrap(sdkerrors.ErrLogic, "account keeper is required for post builder")
 	}
