+++ app/upgrades/v18/constants.go
@@ -1,6 +1,10 @@
 package v18
 
 import (
+	feemarkettypes "github.com/skip-mev/feemarket/x/feemarket/types"
+
+	store "github.com/cosmos/cosmos-sdk/store/types"
+
 	"github.com/cosmos/gaia/v18/app/upgrades"
 )
 
@@ -12,4 +16,9 @@ const (
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
+	StoreUpgrades: store.StoreUpgrades{
+		Added: []string{
+			feemarkettypes.ModuleName,
+		},
+	},
 }
