+++ tests/e2e/e2e_exec_test.go
@@ -323,56 +323,6 @@ func (s *IntegrationTestSuite) execBankMultiSend(
 	s.executeGaiaTxCommand(ctx, c, gaiaCommand, valIdx, s.expectErrExecValidation(c, valIdx, expectErr))
 }
 
-type txBankSend struct {
-	from      string
-	to        string
-	amt       string
-	fees      string
-	log       string
-	expectErr bool
-}
-
-func (s *IntegrationTestSuite) execBankSendBatch(
-	c *chain,
-	valIdx int, //nolint:unparam
-	txs ...txBankSend,
-) int {
-	sucessBankSendCount := 0
-
-	for i := range txs {
-		s.T().Logf(txs[i].log)
-
-		s.execBankSend(c, valIdx, txs[i].from, txs[i].to, txs[i].amt, txs[i].fees, txs[i].expectErr)
-		if !txs[i].expectErr {
-			if !txs[i].expectErr {
-				sucessBankSendCount++
-			}
-		}
-	}
-
-	return sucessBankSendCount
-}
-
-func (s *IntegrationTestSuite) execWithdrawAllRewards(c *chain, valIdx int, payee, fees string, expectErr bool) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
-	gaiaCommand := []string{
-		gaiadBinary,
-		txCommand,
-		distributiontypes.ModuleName,
-		"withdraw-all-rewards",
-		fmt.Sprintf("--%s=%s", flags.FlagFrom, payee),
-		fmt.Sprintf("--%s=%s", flags.FlagGasPrices, fees),
-		fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
-		"--keyring-backend=test",
-		"--output=json",
-		"-y",
-	}
-
-	s.executeGaiaTxCommand(ctx, c, gaiaCommand, valIdx, s.expectErrExecValidation(c, valIdx, expectErr))
-}
-
 func (s *IntegrationTestSuite) execDistributionFundCommunityPool(c *chain, valIdx int, from, amt, fees string) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
 	defer cancel()
@@ -715,7 +665,7 @@ func (s *IntegrationTestSuite) executeGaiaTxCommand(ctx context.Context, c *chai
 	}
 }
 
-func (s *IntegrationTestSuite) executeHermesCommand(ctx context.Context, hermesCmd []string) ([]byte, error) {
+func (s *IntegrationTestSuite) executeHermesCommand(ctx context.Context, hermesCmd []string) ([]byte, error) { //nolint:unparam
 	var outBuf bytes.Buffer
 	exec, err := s.dkrPool.Client.CreateExec(docker.CreateExecOptions{
 		Context:      ctx,
